+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.1
+
+- Added an example of streaming parsing of 1 000 000 rows.
+
+## 0.3.0
+
+- Added experimental support for event-based streaming parsing.
+
 ## 0.2.0
 
 - The parsers have been recreated using PEG grammars.